ingress-adapter-adb-masterdata/requirements.txt
osiris-sdk
pandas

setuptools~=58.0.2
pyodbc~=4.0.32